chore(orders): performance and accessibility enhancements
The OrderItemHeader component now uses a darker shade of green for the text
color when the order is delivered. The OrderItemProduct component now uses
the BlurredImage component instead of NextImage for displaying product images,
with an option to prioritize loading the image. Headings are now in the
ascending order of importance.